pyproject.toml
[project]
name = "explore-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "scikit-learn>=1.6.1",
    "xgboost>=3.0.2",
]
